# src/main/java/com/home/assignment/userapi/dto/UserDTO.java
package com.home.assignment.userapi.dto;

public interface UserDTO {

     Long getUserId();
     String getUserName();
     Long getUnitId();
     String getUnitName();
     String getRoleName();
     Long getRoleId();



}